pyproject.toml
[project]
name = "vidgen"
version = "0.1.0"
description = "Open-source text-to-video MVP service scaffold."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.4.0",
    "fastapi>=0.115.0",
    "numpy>=2.1.0",
    "pillow>=11.0.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "redis>=5.2.0",
    "sqlmodel>=0.0.22",
    "structlog>=24.4.0",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
ml = [
    "accelerate>=1.1.0",
    "controlnet-aux>=0.0.9",
    "diffusers>=0.31.0",
    "imageio[ffmpeg]>=2.36.0",
    "peft>=0.13.0",
    "safetensors>=0.4.5",
    "torch>=2.5.0",
    "transformers>=4.46.0",
]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
